[Fix]가격, 수용인원 수정 [Mark] 결제하기 기능 추가 전

+++ src/main/java/org/pgm/reservationback/dto/RoomDTO.java
@@ -15,7 +15,7 @@ public class RoomDTO {
     private Long id;
     private String name;
     private String description;
-    private int capacity;
-    private Double price;
+    private Long capacity;
+    private Long price;
     private LocalDateTime createdAt;
 }
